# Conflicts:
#	Android/library/runtime/src/main/AndroidManifest.xml
#	CMakeLists.txt
#	Changes.txt
#	Common/ac/common_defines.h
#	Common/ac/gamesetupstruct.cpp
#	Common/ac/gamesetupstruct.h
#	Common/ac/gamesetupstructbase.cpp
#	Common/ac/gamesetupstructbase.h
#	Common/core/def_version.h
#	Common/game/main_game_file.cpp
#	Common/gui/guimain.h
#	Editor/AGS.Editor/AGSEditor.csproj
#	Editor/AGS.Editor/Entities/EditorPreferences.cs
#	Editor/AGS.Editor/GUI/frmMain.cs
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Engine/ac/draw.cpp
#	Engine/ac/draw.h
#	Engine/ac/dynobj/cc_dynamicarray.cpp
#	Engine/ac/dynobj/cc_dynamicarray.h
#	Engine/ac/dynobj/managedobjectpool.cpp
#	Engine/ac/dynobj/managedobjectpool.h
#	Engine/ac/dynobj/scriptuserobject.cpp
#	Engine/ac/dynobj/scriptuserobject.h
#	Engine/ac/room.cpp
#	Engine/ac/roomobject.h
#	Engine/game/savegame_v321.cpp
#	Engine/gui/guidialog.cpp
#	Engine/plugin/agsplugin.cpp
#	Engine/script/cc_instance.cpp
#	OSX/buildlibs/checkenv.sh
#	version.json

+++ Common/font/wfnfont.h
@@ -81,7 +81,7 @@ public:
     }
 
     // Get WFN character for the given code; if the character is missing, returns empty character
-    inline const WFNChar &GetChar(uint8_t code) const
+    inline const WFNChar &GetChar(uint16_t code) const
     {
         return code < _refs.size() ? *_refs[code] : _emptyChar;
     }
